+++ src/client/classes/ac_companion_options.class.js
@@ -6,7 +6,7 @@
 
 import _ from 'lodash';
 
-import { AccountsConf } from 'meteor/pwix:accounts-conf';
+import { AccountsHub } from 'meteor/pwix:accounts-hub';
 import { Options } from 'meteor/pwix:options';
 
 import '../../common/js/index.js';
@@ -36,9 +36,9 @@ export class acCompanionOptions extends Options.Base {
 
     // fields management
     static Fields = [
-        AccountsConf.C.Identifier.NONE,
-        AccountsConf.C.Identifier.OPTIONAL,
-        AccountsConf.C.Identifier.MANDATORY
+        AccountsHub.C.Identifier.NONE,
+        AccountsHub.C.Identifier.OPTIONAL,
+        AccountsHub.C.Identifier.MANDATORY
     ];
 
     // the known render modes