+++ src/client/components/acUserLogin/acUserLogin.js
@@ -7,7 +7,7 @@
 
 import _ from 'lodash';
 
-import { AccountsConf } from 'meteor/pwix:accounts-conf';
+import { AccountsHub } from 'meteor/pwix:accounts-hub';
 import { Random } from 'meteor/random';
 import { ReactiveVar } from 'meteor/reactive-var';
 
@@ -206,10 +206,10 @@ Template.acUserLogin.events({
             case AccountsUI.C.Panel.SIGNUP:
                 if( instance.AC.options.signupSubmit()){
                     let options = {};
-                    if( instance.AC.options.signupHaveUsername() !== AccountsConf.C.Identifier.NONE ){
+                    if( instance.AC.options.signupHaveUsername() !== AccountsHub.C.Identifier.NONE ){
                         options.username = $( '.ac-signup .ac-input-username .ac-input' ).val().trim();
                     }
-                    if( instance.AC.options.signupHaveEmailAddress() !== AccountsConf.C.Identifier.NONE ){
+                    if( instance.AC.options.signupHaveEmailAddress() !== AccountsHub.C.Identifier.NONE ){
                         options.email = $( '.ac-signup .ac-input-email .ac-input' ).val().trim();
                     }
                     options.password = $( '.ac-signup .ac-newone .ac-input' ).val().trim();