+++ src/client/components/ac_signup/ac_signup.js
@@ -5,7 +5,7 @@
  *  - AC: the acUserLogin internal data structure
  */
 
-import { AccountsConf } from 'meteor/pwix:accounts-conf';
+import { AccountsHub } from 'meteor/pwix:accounts-hub';
 import { ReactiveVar } from 'meteor/reactive-var';
 
 import '../ac_input_email/ac_input_email.js';
@@ -61,13 +61,13 @@ Template.ac_signup.onCreated( function(){
                 self.$( '.ac-twice-passwords-sub' ).trigger( 'ac-check' );
             }
             // if an email is mandatory, it must be set here
-            if( self.AC.haveEmailAddress() !== AccountsConf.C.Identifier.NONE ){
+            if( self.AC.haveEmailAddress() !== AccountsHub.C.Identifier.NONE ){
                 if( !self.AC.emailOk.get()){
                     self.$( '.ac-input-email-sub' ).trigger( 'ac-check' );
                 }
             }
             // if a username is mandatory, it must be set here
-            if( self.AC.haveUsername() !== AccountsConf.C.Identifier.NONE ){
+            if( self.AC.haveUsername() !== AccountsHub.C.Identifier.NONE ){
                 if( !self.AC.usernameOk.get()){
                     self.$( '.ac-input-username-sub' ).trigger( 'ac-check' );
                 }
@@ -85,10 +85,10 @@ Template.ac_signup.onCreated( function(){
             return data.ok;
         },
         haveEmailAddress(){
-            return Template.currentData().AC.options.signupHaveEmailAddress() !== AccountsConf.C.Identifier.NONE;
+            return Template.currentData().AC.options.signupHaveEmailAddress() !== AccountsHub.C.Identifier.NONE;
         },
         haveUsername(){
-            return Template.currentData().AC.options.signupHaveUsername() !== AccountsConf.C.Identifier.NONE;
+            return Template.currentData().AC.options.signupHaveUsername() !== AccountsHub.C.Identifier.NONE;
         },
         clearPanel(){
             self.$( 'input' )
@@ -98,8 +98,8 @@ Template.ac_signup.onCreated( function(){
     };
 
     // setup initial default values so that field which is not present is always true
-    self.AC.emailOk.set( Template.currentData().AC.options.signupHaveEmailAddress() !== AccountsConf.C.Identifier.MANDATORY );
-    self.AC.usernameOk.set( Template.currentData().AC.options.signupHaveUsername() !== AccountsConf.C.Identifier.MANDATORY);
+    self.AC.emailOk.set( Template.currentData().AC.options.signupHaveEmailAddress() !== AccountsHub.C.Identifier.MANDATORY );
+    self.AC.usernameOk.set( Template.currentData().AC.options.signupHaveUsername() !== AccountsHub.C.Identifier.MANDATORY);
 });
 
 Template.ac_signup.onRendered( function(){
@@ -142,7 +142,7 @@ Template.ac_signup.helpers({
             withErrorMsg: true,
             withFieldset: this.AC.options.signupFieldset(),
             legend: this.AC.options.signupLegendEmail(),
-            withMandatoryField: this.AC.options.signupHaveEmailAddress() === AccountsConf.C.Identifier.MANDATORY,
+            withMandatoryField: this.AC.options.signupHaveEmailAddress() === AccountsHub.C.Identifier.MANDATORY,
             placeholder: this.AC.options.signupEmailPlaceholder()
         };
     },
@@ -178,7 +178,7 @@ Template.ac_signup.helpers({
             withErrorMsg: true,
             withFieldset: this.AC.options.signupFieldset(),
             legend: this.AC.options.signupLegendUsername(),
-            withMandatoryField: this.AC.options.signupHaveUsername() === AccountsConf.C.Identifier.MANDATORY
+            withMandatoryField: this.AC.options.signupHaveUsername() === AccountsHub.C.Identifier.MANDATORY
         };
     },